Refactor votes row into new component
